requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
async-timeout==5.0.1
attrs==25.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.13.3
bertopic==0.17.0
bitsandbytes==0.45.5
blake3==1.0.4
cachetools==5.5.2
cbor==1.0.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
chest==0.2.3
click==8.1.8
cloudpickle==3.1.1
colaburl==0.1.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
compressed-tensors==0.9.2
contourpy==1.3.0
cupy-cuda12x==13.4.1
cycler==0.12.1
data==0.4
datasets==3.5.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1744321239219/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
faiss-gpu==1.7.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
FlagEmbedding==1.3.4
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
funcsigs==1.0.2
future==1.0.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.32
h11==0.14.0
hdbscan==0.8.40
HeapDict==1.0.1
hf-xet==1.0.3
hf_transfer==0.1.9
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
ijson==3.3.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inscriptis==2.6.0
interegular==0.3.3
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
ir_datasets==0.5.10
ir_measures==0.3.7
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
keybert==0.9.0
kiwisolver==1.4.7
lark==1.2.2
latex==0.7.0
llguidance==0.7.14
llvmlite==0.43.0
lm-format-enforcer==0.10.11
lxml==5.3.2
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matchpy==0.5.5
matplotlib==3.7.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mistral_common==1.5.4
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
multiset==2.1.1
nanobind==2.6.1
narwhals==1.34.1
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
nltk==3.9.1
npids==0.0.2
nptyping==1.4.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
openai==1.73.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
partial-json-parser==0.2.1.1.post5
patsy==1.0.1
peft==0.15.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.2.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1742485085/work
plotly==6.0.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycountry==24.6.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyjnius==1.6.1
pynndescent==0.5.13
pyparsing==3.2.3
pyterrier-adaptive==0.2.0
pyterrier-alpha==0.14.0
pyterrier-doc2query @ git+https://github.com/terrierteam/pyterrier_doc2query.git@1b0b8c190a58836e3c41d375a668c39e9ec11037
pyterrier-dr @ git+https://github.com/terrierteam/pyterrier_dr.git@6a7770ba3e94c91e4e8a123a78bfa3518f392320
pyterrier-pisa==0.0.6
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-terrier==0.13.0
pytrec_eval-terrier==0.5.7
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rank-bm25==0.2.2
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.25.1
setproctitle==1.3.5
shellingham==1.5.4
shtab==1.7.1
shutilwhich==1.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
tempdir==0.7.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615921868/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.51.2
trec-car-tools==2.6
triton==3.2.0
trl==0.17.0
typeguard==4.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
typish==1.9.3
tyro==0.9.18
tzdata==2025.2
umap-learn==0.5.7
unlzw3==0.2.3
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.8.3
wandb==0.19.11
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.0.5
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websockets==15.0.1
wget==3.2
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.17
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zlib-state==0.1.9
